apps/backend/python-worker/requirements.txt
redis==4.5.4
python-dotenv==1.0.0